Move hexadecimal validator to a new module specifically for validators
